requirements.txt
apns2==0.7.2
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-filter==23.2
django-import-export==3.3.1
django-push-notifications==3.0.0
django-sslserver==0.22
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
et-xmlfile==1.1.0
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
py-vapid==1.9.0
pycparser==2.21
PyJWT==1.7.1
pyrsistent==0.19.3
pytz==2023.3
pywebpush==1.14.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
python-telegram-bot==13.14
firebase-admin
